adjust index exports
